# webconsole-back/swagger/SGDSI_Website4sgCore_API_ADMIN_PROD.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Website4sg Core API ADMIN PROD documentation",
    "version" : "1.3",
    "title" : "SGDSI Website4sg Core API ADMIN PROD"
  },
  "host" : "core.website4sg.saint-gobain.com",
  "basePath" : "/",
...
  "schemes" : [ "https" ],
  "paths" : {
    "/tools/incapsula/sites" : {
      "post" : {
        "tags" : [ "toolsincapsulasites" ],
        "summary" : "Add a site in Incapsula",
        "description" : "Add a site in Incapsula",
        "operationId" : "addWebsite",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "Site",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/WAF Site"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Operation successfully done.",
            "schema" : {
              "$ref" : "#/definitions/IncapsulaResponse"
            }
          },
          "400" : {
            "description" : "Invalid data supplied"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "jwt" : [ ]
        } ]
      }
    },
...
    "/tools/incapsula/sites/{domainCode}/cache/mode" : {
      "put" : {
        "tags" : [ "toolsincapsulasites" ],
        "summary" : "Configure cache mode of a site in Incapsula",
        "description" : "Configure cache mode of a site in Incapsula",
        "operationId" : "configureCacheMode",
        "parameters" : [ {
          "name" : "domainCode",
          "in" : "path",
          "description" : "Domain Id",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "Cache Mode",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/WAF CacheMode"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Operation successfully done.",
            "schema" : {
              "$ref" : "#/definitions/IncapsulaResponse"
            }
          },
          "400" : {
            "description" : "Invalid data supplied"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "jwt" : [ ]
        } ]
      }
    },
    "/tools/incapsula/sites/{domainCode}/cache/purge" : {
      "delete" : {
        "tags" : [ "toolsincapsulasites" ],
        "summary" : "Purge cache of a site in Incapsula",
        "description" : "Purge cache of a site in Incapsula",
        "operationId" : "purgeCache",
        "parameters" : [ {
          "name" : "domainCode",
          "in" : "path",
          "description" : "Domain Id",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Operation successfully done.",
            "schema" : {
              "$ref" : "#/definitions/IncapsulaResponse"
            }
          },
          "400" : {
            "description" : "Invalid data supplied"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "jwt" : [ ]
        } ]
      }
    },
    "/tools/incapsula/sites/{domainCode}/cache/rules" : {
      "put" : {
        "tags" : [ "toolsincapsulasites" ],
        "summary" : "Configure cache rules of a site in Incapsula",
        "description" : "Configure cache rules of a site in Incapsula",
        "operationId" : "configureCacheRules",
        "parameters" : [ {
          "name" : "domainCode",
          "in" : "path",
          "description" : "Domain Id",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "Cache Rules",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/WAF CacheRules"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Operation successfully done.",
            "schema" : {
              "$ref" : "#/definitions/IncapsulaResponse"
            }
          },
          "400" : {
            "description" : "Invalid data supplied"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "jwt" : [ ]
        } ]
      }
    },
    "/tools/incapsula/sites/{domainCode}/cache/settings" : {
      "put" : {
        "tags" : [ "toolsincapsulasites" ],
        "summary" : "Configure cache settings of a site in Incapsula",
        "description" : "Configure cache settings of a site in Incapsula",
        "operationId" : "configureCacheSettings",
        "parameters" : [ {
          "name" : "domainCode",
          "in" : "path",
          "description" : "Domain Id",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "Cache Settings",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/WAF CacheSettings"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Operation successfully done.",
            "schema" : {
              "$ref" : "#/definitions/IncapsulaResponse"
            }
          },
          "400" : {
            "description" : "Invalid data supplied"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "jwt" : [ ]
        } ]
      }
    },
...
  },
  "definitions" : {
...
    "IncapsulaResponse" : {
      "type" : "object",
      "properties" : {
        "res" : {
          "type" : "integer",
          "format" : "int32"
        },
        "res_message" : {
          "type" : "string"
        },
        "status" : {
          "type" : "string"
        },
        "ips" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "dns" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/WAF Dn"
          }
        },
        "original_dns" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/WAF OriginalDn"
          }
        },
        "warnings" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/WAF Warning"
          }
        },
        "security" : {
          "$ref" : "#/definitions/WAF Security"
        },
        "active" : {
          "type" : "string"
        },
        "acceleration_level" : {
          "type" : "string"
        },
        "site_creation_date" : {
          "type" : "integer",
          "format" : "int64"
        },
        "sealLocation" : {
          "$ref" : "#/definitions/WAF SealLocation"
        },
        "ssl" : {
          "$ref" : "#/definitions/WAF Ssl"
        },
        "login_protect" : {
          "$ref" : "#/definitions/WAF LoginProtect"
        },
        "performance_configuration" : {
          "$ref" : "#/definitions/WAF PerformanceConfiguration"
        },
        "visits_timeseries" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/WAF VisitsTimeseries"
          }
        },
        "requests_geo_dist_summary" : {
          "$ref" : "#/definitions/WAF RequestsGeoDistSummary"
        },
        "caching" : {
          "$ref" : "#/definitions/WAF Caching"
        },
        "caching_timeseries" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/WAF CachingTimeseries"
          }
        },
        "hits_timeseries" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/WAF HitsTimeseries"
          }
        },
        "threats" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/WAF Threat"
          }
        },
        "visits_dist_summary" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/WAF VisitsDistSummary"
          }
        },
        "bandwidth_timeseries" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/WAF BandwidthTimeseries"
          }
        },
        "visits" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/WAF Visit"
          }
        }
      },
      "description" : "IncapsulaResponse Object"
    },
...
    "WAF" : {
      "type" : "object",
      "properties" : {
        "rules" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/WAF Rules"
          }
        }
      },
      "description" : "WAF Object"
    },
...
  }
}